+++ CHANGES.rst
@@ -6,13 +6,19 @@
  Unreleased
 ************
 
+**Fixed**
+
+- Fixed handling of 'changes' directives (``deprecated``, ``versionadded``,
+  ``versionchanged``, and ``versionremoved``) and will now format the description text
+  within the content body of the directive.
+
 ********************
  2.0.2 (2026/02/07)
 ********************
 
 **Fixed**
 
-- Fix resolving section adornments from pyproject.toml.
+- Fixed resolving section adornments from pyproject.toml.
 
 ********************
  2.0.1 (2025/12/27)